# poezii-api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Poezii API",
    "description": "\n            API cultural dedicat literaturii române - poezii, autori, texte integrale.\n\n            ## Autentificare\n            Pentru a accesa API-ul, includeți header-ul `x-api-key` cu o cheie validă.\n\n            ## Planuri\n            - **FREE**: 100 de cereri pe minut, fără acces la textul integral al poeziilor\n            - **PRO**: 1000 de cereri pe minut, acces complet la toate resursele, inclusiv texte integrale\n            - **ENTERPRISE**: 10000 de cereri pe minut, suport dedicat și SLA \n\n            ## Rate Limiting\n            - FREE: 100 requests / minute\n            - PRO: 1000 requests / minute\n            - ENTERPRISE: 10000 requests / minute\n\n            ## Coduri de răspuns comune\n            - `200 OK` – cerere procesată cu succes\n            - `201 Created` – resursă creată cu succes\n            - `204 No Content` – resursă ștearsă cu succes\n            - `400 Bad Request` – date de intrare invalide\n            - `401 Unauthorized` – cheie API lipsă sau invalidă\n            - `403 Forbidden` – acces interzis (plan insuficient)\n            - `404 Not Found` – resursă negăsită\n            - `429 Too Many Requests` – limită de cereri depășită\n            - `500 Internal Server Error` – eroare internă\n        ",
    "version": "1.0.0",
    "contact": {
      "name": "Poezii API",
      "email": "support@poezii-api.ro"
    },
    "license": {
...
    }
...
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "name": "x-api-key",
        "in": "header",
        "description": "Cheia API obținută de la administrator"
      }
    },
...
  },
  "paths": {
...
    "/authors": {
      "get": {
        "summary": "Lista poeților",
        "tags": [
          "Authors"
        ],
        "description": "Obține o listă paginată de poeți cu căutare și filtrare după secol.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "search",
            "required": false,
            "description": "Caută în numele poetului"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 21
            },
            "in": "query",
            "name": "century",
            "required": false,
            "description": "Filtrare după secol (ex: 19)"
          },
          {
            "schema": {
              "type": "integer",
              "default": 1
            },
            "in": "query",
            "name": "page",
            "required": false,
            "description": "Numărul paginii"
          },
          {
            "schema": {
              "type": "integer",
              "default": 10,
              "maximum": 100
            },
            "in": "query",
            "name": "limit",
            "required": false,
            "description": "Rezultate pe pagină"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "name",
                "birthYear",
                "deathYear"
              ],
              "default": "name"
            },
            "in": "query",
            "name": "sortBy",
            "required": false,
            "description": "Câmp pentru sortare"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "asc"
            },
            "in": "query",
            "name": "sortOrder",
            "required": false,
            "description": "Direcția de sortare"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de poeți returnată cu succes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Lista de poeți returnată cu succes",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "name": {
                            "type": "string"
                          },
                          "slug": {
                            "type": "string"
                          },
                          "birthYear": {
                            "type": "integer"
                          },
                          "deathYear": {
                            "type": "integer"
                          },
                          "popularity": {
                            "type": "integer"
                          }
                        }
                      }
                    },
                    "count": {
                      "type": "integer"
                    },
                    "page": {
                      "type": "integer"
                    },
                    "limit": {
                      "type": "integer"
                    },
                    "_links": {
                      "type": "object",
                      "properties": {
                        "self": {
                          "type": "object",
                          "properties": {
                            "href": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/authors/{id}": {
      "get": {
        "summary": "Detaliile unui poet (după ID)",
        "tags": [
          "Authors"
        ],
        "description": "Obține informațiile complete ale unui poet folosind ID-ul numeric.",
        "parameters": [
          {
            "schema": {
              "type": "integer"
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "ID-ul numeric al poetului"
          }
        ],
        "responses": {
          "200": {
            "description": "Detaliile poetului",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Detaliile poetului",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "name": {
                      "type": "string"
                    },
                    "slug": {
                      "type": "string"
                    },
                    "birthYear": {
                      "type": "integer"
                    },
                    "deathYear": {
                      "type": "integer"
                    },
                    "popularity": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Poetul nu a fost găsit",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Poetul nu a fost găsit",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
...
  "security": [],
...
}